requirements.txt
altgraph==0.17.4
black==25.1.0
blinker==1.9.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
defusedxml==0.7.1
edk==0.390.4
flask==3.1.2
flask-cors==6.0.1
flexivrdk==1.7.0
fonttools==4.59.1
fpdf2==2.8.5
graphviz==0.21
grpcio==1.44.0
grpcio-tools==1.44.0
itsdangerous==2.2.0
jinja2==3.1.6
kiwisolver==1.4.9
loguru==0.7.3
lupa==2.4
markupsafe==3.0.3
matplotlib==3.10.5
mypy-extensions==1.1.0
narwhals==2.12.0
numpy==2.2.6
packaging==24.2
pandas==2.3.3
pandas-stubs==2.3.2.250926
pathspec==0.12.1
pefile==2023.2.7
pillow==11.3.0
pip==25.2
platformdirs==4.3.7
plotly==6.5.0
ppdk==1.3100.1
protobuf==3.20.0
pyarmor==9.0.6
pyarmor-cli-core==7.6.6
pybind11==3.0.0
pybuilder==0.13.17
pycparser==2.22
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.3
pynng==0.8.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32-ctypes==0.2.3
pywebview>=4.0
pyyaml==6.0.2
scipy==1.15.3
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
spdlog==2.0.6
tomli==2.2.1
transitions==0.9.3
typing-extensions==4.13.2
tzdata==2025.2
werkzeug==3.1.3
wheel==0.41.2
win32-setctime==1.2.0
xlsxwriter==3.2.9